[ joleksia ] Advanced collision detection
Now ball uses the more advanced collision detection that allows for horizontal and vertical bounces. It also has a tolerance region if we go into the paddle (if it's in the tolerance region, it will bounce and work as it should, if not, then it'll go pass through the paddle and well, work as it should). Also, the paddles were slightly resized and I've started messing around with the velocity