[ joleksia ] AI and Scoring
Working on AI proceeds making it work more efficient and accurate. Also, there's now a way to score a point if the ball hits the wall. If either player0 or player1 exceeds the score of 10, the matchOver event is sent. There's quite a few things to be done here, check TODO's