[ joleksia ] GUI WIP and ball fix
GUI is in a general 'intended' state; we can iterate from it making it look better or just keep it like that. Also, there's a temporary fix for the ball: setTimeout and starting in Ball::reset() method was commented-out because it caused the issue with game states (the ball was resetting and randomizing it's velocity after a second of match). This change requires an approval of original implementer (supposedly: @agarbacz).